+++ apps/chat/src/components/Marketplace/ApplicationDetails/ApplicationDetails.tsx
@@ -29,7 +29,6 @@ interface Props {
   onDelete: (entity: DialAIEntityModel) => void;
   onChangeVersion: (entity: DialAIEntityModel) => void;
   onBookmarkClick: (entity: DialAIEntityModel) => void;
-  onLogsClick: (entity: string) => void;
 }
 
 const ApplicationDetails = ({
@@ -43,7 +42,6 @@ const ApplicationDetails = ({
   onDelete,
   onChangeVersion,
   onBookmarkClick,
-  onLogsClick,
 }: Props) => {
   const dispatch = useAppDispatch();
 
@@ -96,7 +94,6 @@ const ApplicationDetails = ({
         onEdit={onEdit}
         onDelete={onDelete}
         onBookmarkClick={onBookmarkClick}
-        onLogsClick={onLogsClick}
       />
     </Modal>
   );
