+++ apps/chat/src/components/Marketplace/TabRenderer.tsx
@@ -43,7 +43,6 @@ import { SearchHeader } from '@/src/components/Marketplace/SearchHeader';
 
 import Magnifier from '../../../public/images/icons/search-alt.svg';
 import { NoResultsFound } from '../Common/NoResultsFound';
-import { ApplicationLogs } from './ApplicationLogs';
 
 import { PublishActions, ShareEntity } from '@epam/ai-dial-shared';
 import intersection from 'lodash-es/intersection';
@@ -237,7 +236,6 @@ export const TabRenderer = ({ screenState }: TabRendererProps) => {
     action: PublishActions;
   }>();
   const [detailsModelReference, setDetailsModelReference] = useState<string>();
-  const [isOpenLogs, setIsOpenLogs] = useState<boolean>();
 
   const isSomeFilterNotEmpty =
     searchTerm.length ||
@@ -383,11 +381,6 @@ export const TabRenderer = ({ screenState }: TabRendererProps) => {
     [setApplicationModel],
   );
 
-  const handleCloseApplicationLogs = useCallback(
-    () => setIsOpenLogs(false),
-    [setIsOpenLogs],
-  );
-
   const handleCloseDetailsDialog = useCallback(
     () => setDetailsModelReference(undefined),
     [setDetailsModelReference],
@@ -407,14 +400,6 @@ export const TabRenderer = ({ screenState }: TabRendererProps) => {
       }
     },
     [dispatch, installedModelIds],
-  );
-
-  const handleLogClick = useCallback(
-    (entityId: string) => {
-      dispatch(ApplicationActions.getLogs(entityId));
-      setIsOpenLogs(true);
-    },
-    [dispatch],
   );
 
   const detailsModel = detailsModelReference
@@ -454,14 +439,6 @@ export const TabRenderer = ({ screenState }: TabRendererProps) => {
           type={applicationModel.type}
         />
       )}
-      {isOpenLogs && detailsModel && (
-        <ApplicationLogs
-          isOpen={isOpenLogs}
-          onClose={handleCloseApplicationLogs}
-          entity={detailsModel}
-          onLogsClick={handleLogClick}
-        />
-      )}
       {!!deleteModel && (
         <ConfirmDialog
           isOpen={!!deleteModel}
@@ -480,7 +457,6 @@ export const TabRenderer = ({ screenState }: TabRendererProps) => {
           onDelete={handleDelete}
           onEdit={handleEditApplication}
           onBookmarkClick={handleBookmarkClick}
-          onLogsClick={handleLogClick}
           allEntities={allModels}
           isMyAppsTab={selectedTab === MarketplaceTabs.MY_APPLICATIONS}
         />
